Add Endless Scrolling, Next Page Load Failure, Last Page Limit, Improve loading failure handling
Add EndlessRecyclerViewScrollListener. It handles page counting
internally.

Additionally, if a page fails to load, we need to notify this scroll
listener, so that it can update its internal page count and loading
boolean.

Also, add total pages limit to endless scrolling,, so that onLoadMore()
is not called when the lastPageIndex has been reached.